add scripts together with README

+++ scripts/analyze_app_log.sh
@@ -2,7 +2,7 @@
 
 g_current_dir=$(dirname $(readlink -f $0))
 g_timestamp=`date +%Y%m%m%H%M%S`
-g_tmp_dir=/tmp/sparkevent-${g_timestamp}
+g_tmp_dir=/tmp/mysparkevent-${g_timestamp}
 
 function extract_first_sql_execution() {
   local input_file=$1
@@ -25,6 +25,14 @@ function generate_execution_name() {
   echo "execId${execution_id}-$desc"
 }
 
+function generate_execution_name_caleb() {
+  local input_file=$1
+  local execution_id desc
+  execution_id=`grep "org.apache.spark.sql.execution.ui.SparkListenerSQLExecutionStart" $input_file|jq .executionId`
+  desc=`grep "org.apache.spark.sql.execution.ui.SparkListenerSQLExecutionStart" $input_file|jq .description|tr -d '"'|awk '{print $1}'|tr -d '-'`
+  echo "execId${execution_id}-sql$desc"
+}
+
 function main_log_parser() {
   local tmp_dir=${g_tmp_dir}
   if [ ! -d $tmp_dir ];then
@@ -32,6 +40,7 @@ function main_log_parser() {
   fi
   local input_file=$1
   local output_dir=$2
+  local gen_name=$3
   local jar_file=${g_current_dir}/../target/SparkEventLogAnalytic-1.0-SNAPSHOT-jar-with-dependencies.jar
   local prefix=sparkeventlog
   local tmp_execute_log=$tmp_dir/single_eventlog.txt
@@ -41,15 +50,26 @@ function main_log_parser() {
   then
     mkdir $output_dir
   fi
-
+  if [ ! -e $jar_file ];
+  then
+     cd ${g_current_dir}/../
+     mvn clean package
+     cd -
+  fi
   while [ 1 ]; do
     extract_first_sql_execution $input_file $tmp_execute_log
     if [ ! -s $tmp_execute_log ]; then
       # no execution was found in the log file
       break
     fi
-    local name=$(generate_execution_name $tmp_execute_log)
-    java -Dspark.master=local -jar $jar_file -i $tmp_execute_log -o $output_dir/$name -t -s
+    local f=`grep "Stage Info" ${tmp_execute_log}|head -n1`
+    if [ "$f" == "" ];then
+       echo "no Stage Info: $tmp_execute_log"
+       continue
+    fi
+
+    local name=$($gen_name $tmp_execute_log)
+    java -Dspark.master=local -jar $jar_file -i $tmp_execute_log -o $output_dir/$name -t -s -e
     remove_first_matched_execution $input_file $tmp_follow_log
     if [ -e $tmp_input ]; then
       rm $tmp_input
@@ -59,17 +79,37 @@ function main_log_parser() {
   done
 }
 
-if [ $# -lt 1 ]
-then
-   echo "Specify the <event log file> (<output dir>)"
-   echo "The output dir is /tmp/sparkevent-xxx if you did not specify it"
-   exit 1
-fi
+function usage() {
+cat << EOF
+  $0:<options>
+     -h             print help
+     -b             parse benchmark history log: description contains '/*sqlqxxx*/'
+     -c             parse caleb history log: description contains '--12'
+     -i <inputfile> Required. Specify the input json file
+     -o <outdir>    Optional. Specify the output dir. Default is /tmp/sparkeventxxx
+EOF
+  exit 1
+}
 
-input=$1
+input=""
+bench=0
+caleb=0
 outdir=${g_tmp_dir}
-if [ $# -eq 2 ]
-then
-outdir=$2
+while getopts 'bchi:o:' c
+do
+  case $c in
+    b) bench=1;;
+    c) caleb=1;;
+    i) input=$OPTARG;;
+    o) outdir=$OPTARG;;
+    h) usage;;
+  esac
+done
+gen_name="generate_execution_name"
+if [ "$caleb" == "1" ];then
+  gen_name="generate_execution_name_caleb"
 fi
-main_log_parser $input $outdir
+if [ "$input" == "" ];then
+  usage
+fi
+main_log_parser $input $outdir $gen_name
